Added math lib linkage, refactored cmakes
