+++ src/test/sts_test.c
@@ -1,4 +1,4 @@
-#include "test.h"
+#include "sts_test.h"
 
 int mu_tests_run = 0;
 char mu_err[MU_ERR_LEN] = { 0 };
